moved environment to car directory
